// src/app/model/Search.model.ts
export class SearchReport {
    CREATE_TIME_FROM: string;
    CREATE_TIME_TO: string;
	TIME_FROM: string;
    TIME_TO: string;
	OUT_TIME_FROM: string;
    OUT_TIME_TO: string;
    FilterMode: number;
    Keyword: string;
    ReportTypeCode:string;
	ReportTemplateCode:string;
}

export class SearchRangeDate {
    dayfrom: string;
    monthfrom: string;
	yearfrom: string;
    dayto: string;
	monthto: string;
    yearto: string;
}